refactor fileutils
